hashmap: pre-fetch first key's value during insert_robinhood
seems to reduce LLC misses during the swap loop.  if we're lucky
on array indexing we can get as many as 8 of these in one load.
presumably once the first is pre-fetched, the dynamic prefetcher
will handle the rest if we spend time in the loop body to need them